feat(index.html) Add a styles.css link and make changes to the styles.
Change the colour and the background colour of the div element and the sizing of the font to match the desired-outcome.png file.

Change the font colour and background colour of the p element using an internal styles method.

Change the font colour, background colour, and border colour, of the button by using inline method.